Updates
Medium Blog embed:
<div class='sk-ww-medium-publication-feed' data-embed-id='25457443'></div><script src='https://widgets.sociablekit.com/medium-publication-feed/widget.js' async defer></script>
Text and tables for prices (attached)
Navigation icon for prices (attached)
Please add the 4 icons I sent before to the top of the banner, between the logo and the navigation menu button.
Please move the "Hassle-Free Estimate!" Button above the carousel and have it go to the prices page.
Have the carousel cycle automatically and make the arrow white so they are easier to see.
Add buffer space to top of the carousel to ensure text is not trimmed (see attached screenshot).
Change the "Hassle-Free Estimate" button in the card with a calendar to "Book an Appointment" and have it go to the RepairDesk widget
Change the "Professional Repair Services" card to "Hassle-Free Transparent Pricing", add appropriate image (maybe something like Unsplash Money). Card body text: " We know people want an idea how much something will cost ahead of time, without a hassle! We provide typical ranges/prices for some common services." Include a button navigating to the prices page.
Website background: I tried making a desktop background which should tessellate when repeated vertically (attached).